+++ R/job_single.R
@@ -1,6 +1,6 @@
 #' Build a SGE job script
 #'
-#' This function builds a template for a Sun Grid Ending (SGE) job script
+#' This function builds a template for a Son of Grid Ending (SGE) job script
 #' including array jobs. Check this blog post by John Muschelli to learn
 #' more about array jobs:
 #' <https://hopstat.wordpress.com/2013/11/05/array-and-sequential-cluster-jobs/>.